Moved things into new folders, removed redundant imports
