+++ src/components/DeckGenerator.js
@@ -1,9 +1,11 @@
 import { useState } from "react";
+
 import NumOfCardsPicker from "./NumOfCardsPicker";
 import EmojisPicker from "./EmojisPicker";
 import SubmitButton from "./SubmitButton";
 import Deck from "./Deck";
-import { generate_cards } from "../generate_cards";
+
+import { generate_cards } from "../utils/generate_cards";
 
 export default function DeckGenerator() {
   const [numOfCards, setNumOfCards] = useState("(select one)");
