+++ src/components/Header.js
@@ -1,5 +1,3 @@
-import Link from "next/link";
-
 export default function Header() {
   return (
     <header>
