--- src/test.js
@@ -1,26 +1,0 @@
-const { generate_cards } = require("./generate_cards.js");
-
-console.log(
-  generate_cards(13, [
-    "😭",
-    "✨",
-    "❤️",
-    "🔥",
-    "🙏",
-    "🎉",
-    "🌹",
-    "👀",
-    "💋",
-    "😡",
-    "💥",
-    "💯",
-    "☀️",
-    "🍀",
-    "✅",
-    "🌈",
-    "💐",
-    "🌺",
-    "💦",
-    "⭐",
-  ])
-);
